+++ src/onnx/parse_gelu.cpp
@@ -21,92 +21,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-#include <migraphx/onnx/op_parser.hpp>
 #include <migraphx/ranges.hpp>
 #include <migraphx/instruction.hpp>
-#include <migraphx/make_op.hpp>
+#include <migraphx/op/builder/insert.hpp>
+#include <migraphx/onnx/op_parser.hpp>
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
 namespace onnx {
-
-static instruction_ref parse_quick_gelu(const onnx_parser::node_info& info, instruction_ref x)
-{
-    // computes x * sigmoid(alpha * x)
-    auto x_type    = x->get_shape().type();
-    auto alpha_val = info.attributes.at("alpha").f();
-    auto alpha     = info.add_literal(migraphx::literal{migraphx::shape{x_type}, {alpha_val}});
-    auto mul_alpha = info.add_common_op("mul", alpha, x);
-    auto sigmoid   = info.add_instruction(migraphx::make_op("sigmoid"), mul_alpha);
-    return info.add_common_op("mul", x, sigmoid);
-}
-
-static instruction_ref parse_gelu_erf(const onnx_parser::node_info& info, instruction_ref x)
-{
-    auto x_type = x->get_shape().type();
-    auto half   = info.add_literal(migraphx::literal{migraphx::shape{x_type}, {0.5f}});
-    auto one    = info.add_literal(migraphx::literal{migraphx::shape{x_type}, {1.0f}});
-    auto sqrt2 =
-        info.add_literal(migraphx::literal{migraphx::shape{x_type}, {static_cast<float>(M_SQRT2)}});
-    auto mul_half = info.add_common_op("mul", x, half);
-    auto div      = info.add_common_op("div", x, sqrt2);
-    auto erf      = info.add_instruction(migraphx::make_op("erf"), div);
-    auto add_one  = info.add_common_op("add", erf, one);
-    return info.add_common_op("mul", mul_half, add_one);
-}
-
-static instruction_ref
-parse_gelu_tanh(const onnx_parser::node_info& info, instruction_ref x, bool fast)
-{
-    auto x_type        = x->get_shape().type();
-    auto fit_const_val = fast ? 0.035677 : 0.044715;
-    auto fit_const = info.add_literal(migraphx::literal{migraphx::shape{x_type}, {fit_const_val}});
-    auto sqrt_2_rpi_val = fast ? 0.797885 : sqrt(M_2_PI);
-    auto sqrt_2_rpi =
-        info.add_literal(migraphx::literal{migraphx::shape{x_type}, {sqrt_2_rpi_val}});
-    auto one   = info.add_literal(migraphx::literal{migraphx::shape{x_type}, {1.0f}});
-    auto half  = info.add_literal(migraphx::literal{migraphx::shape{x_type}, {0.5f}});
-    auto three = info.add_literal(migraphx::literal{migraphx::shape{x_type}, {3.0f}});
-
-    // [0.044715|0.035677] * x^3
-    auto pow0 = info.add_common_op("pow", x, three);
-    auto mul0 = info.add_common_op("mul", pow0, fit_const);
-    instruction_ref tanh_in;
-    if(fast)
-    {
-        // approx = 0.797885 * x + 0.035677 * x^3
-        auto mul1 = info.add_common_op("mul", sqrt_2_rpi, x);
-        tanh_in   = info.add_common_op("add", mul0, mul1);
-    }
-    else
-    {
-        // approx = sqrt(2/pi) * (x + 0.044715 * x^3
-        auto add0 = info.add_common_op("add", mul0, x);
-        tanh_in   = info.add_common_op("mul", add0, sqrt_2_rpi);
-    }
-
-    // 0.5 * x * (1 + Tanh(approx))
-    auto tanh0 = info.add_instruction(migraphx::make_op("tanh"), tanh_in);
-    auto add1  = info.add_common_op("add", tanh0, one);
-    auto mul2  = info.add_common_op("mul", x, half);
-    return info.add_common_op("mul", add1, mul2);
-}
-
-static instruction_ref parse_split_gelu(const onnx_parser::node_info& info, instruction_ref x)
-{
-    size_t last_dim_size = x->get_shape().lens().back();
-    if(last_dim_size < 2 or last_dim_size % 2 != 0)
-        MIGRAPHX_THROW("PARSE_GELU: BiasSplitGelu must have even last dimension which is >= 2");
-    auto split_left = info.add_instruction(
-        migraphx::make_op("slice",
-                          {{"axes", {-1}}, {"starts", {0}}, {"ends", {last_dim_size / 2}}}),
-        x);
-    auto split_right = info.add_instruction(
-        migraphx::make_op(
-            "slice", {{"axes", {-1}}, {"starts", {last_dim_size / 2}}, {"ends", {last_dim_size}}}),
-        x);
-    return info.add_common_op("mul", split_left, parse_gelu_erf(info, split_right));
-}
 
 struct parse_gelu : op_parser<parse_gelu>
 {
@@ -135,7 +57,8 @@ struct parse_gelu : op_parser<parse_gelu>
 
         if(opd.onnx_name == "QuickGelu")
         {
-            return parse_quick_gelu(info, x);
+            auto alpha = info.attributes.at("alpha").f();
+            return op::builder::add("gelu_quick", *info.mod, {x}, {{"alpha", alpha}}).at(0);
         }
 
         if(opd.onnx_name == "FastGelu")
@@ -165,16 +88,16 @@ struct parse_gelu : op_parser<parse_gelu>
         {
             // add should've been inserted from previous conditional statement
             assert(args.size() == 2);
-            return parse_split_gelu(info, x);
+            return op::builder::add("gelu_split", *info.mod, {x}, {}).at(0);
         }
 
         if(approximate == "tanh")
         {
-            return parse_gelu_tanh(info, x, fast);
+            return op::builder::add("gelu_tanh", *info.mod, {x}, {{"fast", fast}}).at(0);
         }
         else
         {
-            return parse_gelu_erf(info, x);
+            return op::builder::add("gelu_erf", *info.mod, {x}, {}).at(0);
         }
     }
 };
