+++ test/include/pointwise.hpp
@@ -65,7 +65,7 @@ migraphx::instruction_ref add_pointwise(migraphx::program& p,
                                         const std::vector<migraphx::instruction_ref>& inputs,
                                         const F& f)
 {
-    return add_pointwise(p, p.get_main_module(), name, std::move(inputs), std::move(f));
+    return add_pointwise(p, p.get_main_module(), name, inputs, std::move(f));
 }
 
 inline auto noop_pointwise()
